+++ src/modelo/vo/CursaVO.java
@@ -47,7 +47,7 @@ public class CursaVO {
 		return nota;
 	}
 
-	public void setNota(int nota) {
+	public void setNota(float nota) {
 		this.nota = nota;
 	}
 }
